Move fetchAddressesData into Client class
