+++ src/utils/addresses.ts
@@ -19,7 +19,6 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 import * as Clipboard from 'expo-clipboard'
 import Toast from 'react-native-root-toast'
 
-import client from '../api/client'
 import { storeAddressMetadata } from '../storage/wallets'
 import { Address, AddressHash, AddressPartial, AddressSettings } from '../types/addresses'
 
@@ -29,45 +28,6 @@ export const getAddressDisplayName = (address: Address): string =>
 export const copyAddressToClipboard = (addressHash: AddressHash) => {
   Clipboard.setString(addressHash)
   Toast.show('Address copied!')
-}
-
-export const fetchAddressesData = async (addressHashes: AddressHash[]) => {
-  const results = []
-
-  for (const addressHash of addressHashes) {
-    console.log('⬇️ Fetching address details: ', addressHash)
-    const { data } = await client.explorerClient.getAddressDetails(addressHash)
-    const availableBalance = data.balance
-      ? data.lockedBalance
-        ? (BigInt(data.balance) - BigInt(data.lockedBalance)).toString()
-        : data.balance
-      : undefined
-
-    console.log('⬇️ Fetching 1st page of address confirmed transactions: ', addressHash)
-    const { data: transactions } = await client.explorerClient.getAddressTransactions(addressHash, 1)
-
-    console.log('⬇️ Fetching address tokens: ', addressHash)
-    const { data: tokenIds } = await client.explorerClient.addresses.getAddressesAddressTokens(addressHash)
-
-    const tokens = await Promise.all(
-      tokenIds.map((id) =>
-        client.explorerClient.addresses.getAddressesAddressTokensTokenIdBalance(addressHash, id).then(({ data }) => ({
-          id,
-          balances: data
-        }))
-      )
-    )
-
-    results.push({
-      hash: addressHash,
-      details: data,
-      availableBalance: availableBalance,
-      transactions,
-      tokens
-    })
-  }
-
-  return results
 }
 
 export const findNextAvailableAddressIndex = (startIndex: number, skipIndexes: number[] = []) => {
